Add ApiProviderConfig class and update Runner to include AnalyzerAgent
