+++ packages/cli/src/options.ts
@@ -1,10 +1,12 @@
-import os from 'node:os'
 // Generated by polka.codes
+
+import os from 'node:os'
 import { AiServiceProvider } from '@polka-codes/core'
 import type { Command } from 'commander'
 import { set } from 'lodash'
 
-import { type Config, loadConfig } from './config'
+import { ApiProviderConfig } from './ApiProviderConfig'
+import { loadConfig } from './config'
 
 export interface CliOptions {
   config?: string
@@ -26,52 +28,6 @@ export function addSharedOptions(command: Command) {
     .option('--max-messages <iterations>', 'Maximum number of messages to send. Default to 50', Number.parseInt, 50)
     .option('--budget <budget>', 'Budget for the AI service. Default to $1000', Number.parseFloat, 1000)
     .option('-v --verbose', 'Enable verbose output. Use -v for level 1, -vv for level 2', (value, prev) => prev + 1, 0)
-}
-
-export class ApiProviderConfig {
-  readonly defaultProvider?: AiServiceProvider
-  readonly providers: Readonly<Partial<Record<AiServiceProvider, { apiKey: string; defaultModel?: string }>>>
-  readonly commands?: Partial<Record<string, { provider?: AiServiceProvider; model?: string }>>
-  readonly agents?: Partial<Record<string, { provider?: AiServiceProvider; model?: string }>>
-
-  constructor(config: Config) {
-    this.defaultProvider = config.defaultProvider as AiServiceProvider | undefined
-    this.providers = config.providers ?? {}
-    this.commands = config.commands as Partial<Record<string, { provider?: AiServiceProvider; model?: string }>>
-    this.agents = config.agents as Partial<Record<string, { provider?: AiServiceProvider; model?: string }>>
-  }
-
-  getConfigForCommand(command: string) {
-    // TODO: strong type command
-    const { provider, model } = this.commands?.[command] ?? this.commands?.default ?? {}
-    const finalProvider = provider ?? this.defaultProvider
-    if (!finalProvider) {
-      return undefined
-    }
-    const finalModel = model ?? this.providers[finalProvider]?.defaultModel
-    const apiKey = this.providers[finalProvider]?.apiKey
-    return {
-      provider: finalProvider,
-      model: finalModel,
-      apiKey,
-    }
-  }
-
-  getConfigForAgent(agent: string) {
-    // TODO: strong type agent
-    const { provider, model } = this.agents?.[agent] ?? this.agents?.default ?? {}
-    const finalProvider = provider ?? this.defaultProvider
-    if (!finalProvider) {
-      return undefined
-    }
-    const finalModel = model ?? this.providers[finalProvider]?.defaultModel
-    const apiKey = this.providers[finalProvider]?.apiKey
-    return {
-      provider: finalProvider,
-      model: finalModel,
-      apiKey,
-    }
-  }
 }
 
 export function parseOptions(options: CliOptions, cwd: string = process.cwd(), home: string = os.homedir()) {
